Refactored `FolderTreeItem`
